% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{connect_sqlite}
\alias{connect_sqlite}
\title{Create a local sqlite connection}
\usage{
connect_sqlite(auto_disconnect = TRUE)
}
\arguments{
\item{auto_disconnect}{Should the connection be automatically closed when the src is deleted? Set to TRUE if you initialize the connection the call to src_dbi(). Pass NA to auto-disconnect but print a message when this happens. cursor}
}
\description{
Wrapper function to return a sqlite connection cursor for local development.
}
\examples{
connect_sqlite()

}
