% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{create_schema}
\alias{create_schema}
\title{Add schema to a sqlite database}
\usage{
create_schema(schema, con)
}
\arguments{
\item{schema}{Schema name}

\item{con}{A sqlite connection cursor}
}
\description{
Wrapper function to create a new schema in a sqlite database for local development.
}
\examples{
con <- connect_sqlite()
create_schema(con, schema = "demo")

}
