% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{db.read_table}
\alias{db.read_table}
\title{Read remote database tables into data frames with validation}
\usage{
db.read_table(con, table, schema = NA, collect = TRUE,
  error_value = NA)
}
\arguments{
\item{con}{A DBIConnection object, as returned by dbConnect().}

\item{table}{A character string specifying the DBMS table name.}

\item{schema}{A character string specifying the schema in which the table is nested.}

\item{collect}{A boolean specifying whether the query results should be collected into memory or left as a lazy query.}

\item{error_value}{Error value to return if dbReadTable fails. Default is NA.}
}
\description{
Wrapper function to read table from schema, return NA by default if permission error (or other).
}
\examples{
con <- connect_sqlite(auto_disconnect = FALSE)
dplyr::copy_to(con, iris, "df", temporary = FALSE)
db.read_table(con = con, table = 'df')

}
