% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{db.write_table}
\alias{db.write_table}
\title{Read remote database tables into data frames with validation}
\usage{
db.write_table(con, data, table, schema = NA, append_only = F,
  drop_overwrite = NA)
}
\arguments{
\item{con}{A DBIConnection object, as returned by dbConnect().}

\item{data}{A data.frame, tbl, or other valid sql data type containing the data to write to the database.}

\item{table}{A character string specifying the DBMS table name.}

\item{schema}{A character string specifying the schema in which the table is nested.}

\item{append_only}{A boolean specifying whether the operation is INSERT or UPDATE.
Default of append_only = F means execute DELETE on table, and update with new data.}

\item{drop_overwrite}{A boolean specifying whether the operation is DROP and INSERT. This will overwrite any existing field types.}
}
\description{
Wrapper function to write data to table in schema, return TRUE if successful, FALSE otherwise.
}
\examples{
connection <- connect_sqlite()

db.write_table(
con = connection, table = 'iris', data = iris)

}
