% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_notes}
\alias{demo_notes}
\title{Demo notes for testing shinyNote module.}
\format{A tibble with 274 rows and 3 variables:
\describe{
  \item{package}{package title, character class}
  \item{category}{function name, character class}
  \item{update}{function title, character class}
  ...
}}
\source{
\href{https://cran.r-project.org/web/packages/available_packages_by_name.html}{CRAN Package Help Pages}
}
\usage{
demo_notes
}
\description{
A dataset containing package functions and their titles for the shiny, shinyWidgets and dplyr packages.
Formatted in a structure compatible with the shinyNote's shiny module.
}
\keyword{datasets}
