% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emojis}
\alias{emojis}
\title{Demo notes for testing \code{shinynote} module.}
\format{
A named list of length 2 with elements of length 1510:
\describe{
  \item{name}{emoji name, character class}
  \item{url}{emoji image url, character class}
  ...
}
}
\source{
\href{https://api.github.com/emojis}{GitHub emojis API}
}
\usage{
emojis
}
\description{
A dataset containing package functions and their titles for the \code{shiny}, \code{shinyWidgets} and \code{dplyr} packages.
Formatted in a structure compatible with the \code{shinyNotes::shinynotes} module.
}
\keyword{datasets}
