% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markdown_notes}
\alias{markdown_notes}
\title{Demo notes formatted with markdown for testing \code{shinynote} module.}
\format{
A \code{tibble} with 3 rows and 3 variables:
\describe{
  \item{formatting}{text format type, character class}
  \item{category}{type of markdown formatter, character class}
  \item{update}{text with markdown syntax, character class}
  ...
}
}
\usage{
markdown_notes
}
\description{
A dataset containing examples of markdown syntax for including emojis, headers, and code blocks.
Formatted in a structure compatible with the \code{shinyNotes::shinynotes} module.
}
\keyword{datasets}
