% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_crud.R
\name{shinynotes}
\alias{shinynotes}
\title{Shiny Notes module - server function}
\usage{
shinynotes(input, output, session, group_column, selected_group,
  group_options, table_id, db_conn, category_options = NA,
  style_options = default_styles())
}
\arguments{
\item{input}{Standard \code{shiny} input}

\item{output}{Standard \code{shiny} output}

\item{session}{Standard \code{shiny} session}

\item{group_column}{Column in table to group and filter notes by}

\item{selected_group}{Currently selected group column value}

\item{group_options}{Group column row value options}

\item{table_id}{Named list with member 'table' and 'schema' referring to a db table containing notes.}

\item{db_conn}{database connection cursor}

\item{category_options}{Category column row value options. Useful if table is empty. Default is NA (retrieved from data)}

\item{style_options}{Optional named list of css styles to apply to note panel elements.}
}
\description{
Server function for the shinynotes module.
}
\examples{
if(interactive()){
  shiny::callModule(
    module = shinynotes, 
    id = "paragraph",
    style_options = shiny::reactive({
    list(
      "type" = "bullets", 
      "header" = list("color" = "#ccc"),
      "panel" = list("scrollY" = TRUE)
      )
    }),
    group_column = "package",
    selected_group = shiny::reactive("shiny"),
    group_options = c("shiny", "shinyWidgets", "dplyr"),
    table_id = list(table = "scroll_demo", schema = "notes"),
    db_conn = connect_sqlite(auto_disconnect = F)
  )
}

}
