% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_crud.R
\name{shinynotesUI}
\alias{shinynotesUI}
\title{Shiny Notes module - UI function}
\usage{
shinynotesUI(id)
}
\arguments{
\item{id}{An ID string that will be used to assign the module's namespace.}
}
\description{
UI function for the shinynotes module.
}
\examples{
if(interactive()){
 shinynotesUI(id = 'paragraph')
}
}
