% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{create_schema}
\alias{create_schema}
\title{Add schema to a SQLite database}
\usage{
create_schema(schema, con)
}
\arguments{
\item{schema}{Schema name}

\item{con}{A \code{\link[=RSQLite]{SQLiteConnection-class}} object, produced
by \code{\link[=DBI]{dbConnect()}} or \code{shinyNotes::connect_sqlite()}}
}
\value{
None. Executes SQL query and returns silently.
}
\description{
Wrapper function to create a new schema in a SQLite database for
  local development.
}
\examples{
con <- connect_sqlite()
create_schema(con, schema = "demo")

}
