% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_notes}
\alias{demo_notes}
\title{Demo notes for testing \code{shinynote} module.}
\format{
A \code{tibble} with 274 rows and 3 variables:
\describe{
  \item{package}{package title, character class}
  \item{category}{function name, character class}
  \item{update}{function title, character class}
  ...
}
}
\source{
\href{https://CRAN.R-project.org/package=shiny}{shiny help pages}

\href{https://CRAN.R-project.org/package=shinyWidgets}{shinyWidgets help pages}

\href{https://CRAN.R-project.org/package=dplyr}{dplyr help pages}
}
\usage{
demo_notes
}
\description{
A dataset containing package functions and their titles for the \code{shiny}, \code{shinyWidgets} and \code{dplyr} packages.
Formatted in a structure compatible with the \code{shinyNotes::shinynotes} module.
}
\keyword{datasets}
