% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_crud.R
\name{shinynotesUI}
\alias{shinynotesUI}
\title{Shiny notes module - UI function}
\usage{
shinynotesUI(id)
}
\arguments{
\item{id}{An ID string that will be used to assign the module's namespace.}
}
\value{
Note module UI, containing note panel and control buttons. An HTML
  tag object that can be rendered as HTML using
  \code{\link[=base]{as.character()}}.
}
\description{
UI function for the \code{shinynotes} module.
}
\examples{
if(interactive()){
 shinynotesUI(id = 'paragraph')
}

}
